Create LinksAnalyticsService
