+++ src/server/services/links.ts
@@ -1,13 +1,16 @@
-import { Link, LinkAnalytics, PrismaClient } from "@prisma/client";
+import { Link, PrismaClient } from "@prisma/client";
 import prisma from "@/lib/db/prisma";
 import { ModifiedLink } from "@/lib/store/slices/linksSlice";
 import { AnalyticsData } from "@/app/components/types";
+import { LinksAnalyticsService } from "./linksAnalytics";
 
 export class LinksService {
   private readonly prisma: PrismaClient;
+  private readonly linksAnalyticsService: LinksAnalyticsService;
 
   constructor() {
     this.prisma = prisma;
+    this.linksAnalyticsService = new LinksAnalyticsService();
   }
 
   async getLinksByUser(userId: string): Promise<Link[] | null> {
@@ -141,11 +144,11 @@ export class LinksService {
 
     try {
       await Promise.all(links.map(async (link) => {
-        const analyticsData = await this.getAnalyticsByTimespan("7D", link.id);
+        const analyticsData = await this.linksAnalyticsService.getByTimespan("7D", link.id);
 
         const analytics: AnalyticsData = {
           linkName: link.platform,
-          analytics: this.mapAnalyticsData(analyticsData),
+          analytics: this.linksAnalyticsService.mapData(analyticsData),
         };
 
         linksAnalytics.push(analytics);
@@ -156,54 +159,6 @@ export class LinksService {
       console.error(error);
       return null;
     }
-  }
-
-  // TODO: Move to "LinksAnalyticsService"
-  private async getAnalyticsByTimespan(span: "7D", linkId: string): Promise<LinkAnalytics[]> {
-    if (span === "7D") {
-      return await this.getAnalyticsLast7Days(linkId) || [];
-    }
-
-    return [];
-  }
-
-  // TODO: Move to "LinksAnalyticsService"
-  private async getAnalyticsLast7Days(linkId: string): Promise<LinkAnalytics[] | null> {
-    // TODO: Move to "DateUtils"
-    const today = new Date().toISOString().split('T')[0];
-    const sevenDaysAgo = new Date();
-    sevenDaysAgo.setDate(sevenDaysAgo.getDate() - 7);
-    const formattedSevenDaysAgo = sevenDaysAgo.toISOString().split('T')[0]; 
-
-    try {
-      const analytics = await this.prisma.linkAnalytics.findMany({
-        where: {
-          linkId,
-          createdAt: {
-            gte: new Date(formattedSevenDaysAgo + 'T00:00:00.000Z'),
-            lte: new Date(today + 'T23:59:59.999Z'),
-          },
-        },
-        orderBy: {
-          createdAt: 'asc',
-        }
-      });
-
-      return analytics;
-    } catch (error) {
-      console.error(error);
-      return null;
-    }
-  } 
-
-  // TODO: Move to "LinksAnalyticsService"
-  private mapAnalyticsData (analytics: LinkAnalytics[]): { date: Date; clicks: string; }[] {
-    return analytics.map((data) => {
-      return {
-        date: data.createdAt,
-        clicks: data.clicks.toString(),
-      };
-    });
   }
 
   private checkLinksModification(links: ModifiedLink[]): null | "OK" {
